+++ tests/Unit/FactoryOverLeagueTest.php
@@ -4,6 +4,8 @@ namespace Yiisoft\Factory\Tests\Unit;
 
 use Psr\Container\ContainerInterface;
 use League\Container\Container;
+use Yiisoft\Factory\Factory;
+use Yiisoft\Factory\Definitions\Reference;
 
 /**
  * Test the Factory over League Container.
@@ -22,5 +24,26 @@ class FactoryOverLeagueTest extends AbstractFactoryTest
         }
 
         return $container;
+    }
+
+    public function testCreateFactory(): void
+    {
+        $container = $this->createContainer();
+        $this->setupContainer($container, [ContainerInterface::class => $container]);
+        $factory = new Factory($container, [
+            'factory' => [
+                '__class' => Factory::class,
+                '__construct()' => [
+                    'parent'        => Reference::to(ContainerInterface::class),
+                    'definitions'   => [],
+                ],
+            ],
+        ]);
+        $one = $factory->create('factory');
+        $two = $factory->create('factory');
+        $this->assertNotSame($one, $two);
+        $this->assertNotSame($one, $factory);
+        $this->assertInstanceOf(Factory::class, $one);
+        $this->assertInstanceOf(Factory::class, $two);
     }
 }
