+++ tests/Unit/FactoryOverYiisoftTest.php
@@ -4,6 +4,8 @@ namespace Yiisoft\Factory\Tests\Unit;
 
 use Psr\Container\ContainerInterface;
 use Yiisoft\Di\Container;
+use Yiisoft\Factory\Factory;
+use Yiisoft\Factory\Definitions\Reference;
 
 /**
  * Test the Factory over Yiisoft Container.
@@ -12,15 +14,30 @@ class FactoryOverYiisoftTest extends AbstractFactoryTest
 {
     public function createContainer(iterable $definitions = []): ContainerInterface
     {
-        return $this->setupContainer(new Container(), $definitions);
+        return new Container($definitions);
     }
 
-    public function setupContainer(ContainerInterface $container, iterable $definitions = []): ContainerInterface
+    public function testCreateFactory(): void
     {
-        foreach ($definitions as $id => $definition) {
-            $container->set($id, $definition);
-        }
-
-        return $container;
+        $container = $this->createContainer([
+            ContainerInterface::class => static function (ContainerInterface $container) {
+                return $container;
+            },
+        ]);
+        $factory = new Factory($container, [
+            'factory' => [
+                '__class' => Factory::class,
+                '__construct()' => [
+                    'parent'        => Reference::to(ContainerInterface::class),
+                    'definitions'   => [],
+                ],
+            ],
+        ]);
+        $one = $factory->create('factory');
+        $two = $factory->create('factory');
+        $this->assertNotSame($one, $two);
+        $this->assertNotSame($one, $factory);
+        $this->assertInstanceOf(Factory::class, $one);
+        $this->assertInstanceOf(Factory::class, $two);
     }
 }
